Refactor delay duration types
