+++ packages/types/src/shared/index.ts
@@ -1,1 +1,2 @@
 export * from "./typeUtils"
+export * from "./duration"
